Starting to add properties in the new StructureData.
(1) moving from structure.py to property.py all the classes used in the definition of a property in the StructureData node
(2) creation of the magnetic and hubbard properties (in magnetic.py and hubbard.py). Magnetic is still under some development.
(3) The hubbard.py is actually very similar to the aiida-core/src/aiida_quantumespresso/common/hubbard.py file:
    it is the same file but adjusted to have the Hubbard class as BaseProperty, plus some fixing to support the parent
    feature (see property.py) and a consistent way to generate the hubbard property from built-in-methods (like the
    from_list() one).
     - set default value for the input parameters of the Hubbard class (needed to initialize the class correctly as Hubbard and not Property)
     - set_property method added in the return of the from_list method of the Hubbard class
     - the from_list method is no more a static method of the class.
     - adding hubbard_test.py to test that the property can be set by using the from_list method.
(4) For now, the hubbard property is stored in the DB, not in the repository as the old implementation; TO CHANGE
(5) TO ADD the methods that are contained in the HubbardStructureData of the QE plugin; I think this can be done easily, I will do it in the next push