Adding the new implementation by means of the aiida_atomistic/data/structure/new_implementation.ipynb file.
In the following we define the new StructureData class, here named `StructureData_prototype` for simplicity, as it should be implemented in the `aiida-atomistic` package.
We then define the `PropertyCollector` class, which manages all the properties taking care of all data validation and storage instructions. Instance of this class will be created within the initialization of the StructureData instance and is stored under the `properties` attribute.

In the final implementation, original StructureData will be named `LegacyStructureData` and StructureData_prototype `StructureData`.

- properties are stored under the `properties` attribute of the `StructureData` instance. This attribute is immutable. `STATUS: done`;
- the `properties` attribute is a dictionary, keys being the name of each property and values being a dictionary with the information about the property, which we will store when the StructureData node is stored in the AiiDA db. `STATUS: storage part to be done`;
- information about a property can be provided in only one format (we may implement, in each property, some method to translate from common formats to the accepted one). `STATUS: todo the possiblity to have property class methods`;
- no direct change of the properties is allowed. `STATUS: done`;
- backward compatibility: structure.pbc and structure.kinds are still there. `STATUS: done`;
- always possible to access the list of defined and of supported properties. `STATUS: done`;
- support for custom properties. `STATUS: todo`;
- constructors: adapt methods like `from_ase` to support also the interface with the properties stored in a given `Atoms` object. `STATUS: todo`.

The new StructureData class will be a subclass of the old StructureData. This is the best way to ensure
backward compatibility while avoiding code duplication. This enhance also how this new AiiDA data type is indeed and extension of the original StructureData.
In the future, the entire orm.StructureData module will be moved into the `aiida-atomistic` package and the `pbc` and `kinds` attribute will be removed.