Fixing kinds (#75)
* Adding the to_kinds method

this substitutes the generate_kinds method in the GetterMixin class.
The generate_kinds is moved to utils_kinds.py file, and now it is just a python function.
In the same file there is also a to_kinds function, which is the function that
effectively does the conversion. In the to_kinds method of the GetterMixin, it is indeed
called and also optionally dressed by calcfunction(to_kinds), if provenance is required
(i.e. if StructureData is the instance from which we generate kinds).

* Fixing docs for kinds

and also, moving all the stuff about kinds: from utils to utils_kinds